create reusable useEventHistoryState hook
